+++ source/Classes/nxFileSystemMode.ps1
@@ -1,6 +1,6 @@
 class nxFileSystemMode
 {
-    hidden static [string] $SymbolicTriadParser = '^[-dl]?(?<User>[-wrxsStT]{3})(?<Group>[-wrxsStT]{3})(?<Others>[-wrxsStT]{3})$'
+    hidden static [string] $SymbolicTriadParser = '^[-dlsp]?(?<User>[-wrxsStT]{3})(?<Group>[-wrxsStT]{3})(?<Others>[-wrxsStT]{3})$'
     hidden static [string] $SymbolicOperationParser = '^(?<userClass>[ugoa]{1,3})(?<operator>[\-\+\=]{1})(?<permissions>[wrxTtSs-]{1,3})$'
     [nxFileSystemSpecialMode]  $SpecialModeFlags
     [nxFileSystemAccessRight]  $OwnerMode
@@ -221,5 +221,15 @@ class nxFileSystemMode
 
         Write-Verbose -Message "SymbolNotation: $SymbolNotation"
         return ($SymbolNotation -join '')
+    }
+
+    [string] ToOctal()
+    {
+        return ('{0}{1}{2}{3}' -f (
+            ([int]$this.SpecialModeFlags),
+            ([int]$this.OwnerMode),
+            ([int]$this.GroupMode),
+            ([int]$this.OthersMode)
+        ))
     }
 }