+++ source/Enum/nxFileSystemItemType.ps1
@@ -3,4 +3,6 @@ enum nxFileSystemItemType
     File
     Directory
     Link
+    Pipe
+    Socket
 }