+++ source/Public/Get-nxChildItem.ps1
@@ -1,20 +1,26 @@
 function Get-nxChildItem
 {
+    [CmdletBinding(DefaultParameterSetName = 'default')]
+    [OutputType([nxFileSystemInfo[]])]
     param
     (
-        [Parameter()]
+        [Parameter(ParameterSetName = 'default'         , Position = 0, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
+        [Parameter(ParameterSetName = 'FilterDirectory' , Position = 0, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
+        [Parameter(ParameterSetName = 'FilterFile'      , Position = 0, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
         [System.String[]]
         $Path = '.',
 
-        [Parameter()]
+        [Parameter(ParameterSetName = 'default'         , Position = 1, ValueFromPipelineByPropertyName = $true)]
+        [Parameter(ParameterSetName = 'FilterDirectory' , Position = 1, ValueFromPipelineByPropertyName = $true)]
+        [Parameter(ParameterSetName = 'FilterFile'      , Position = 1, ValueFromPipelineByPropertyName = $true)]
         [Switch]
         $Recurse,
 
-        [Parameter()]
+        [Parameter(ParameterSetName = 'FilterDirectory' , ValueFromPipelineByPropertyName = $true, Position = 2)]
         [Switch]
         $Directory,
 
-        [Parameter()]
+        [Parameter(ParameterSetName = 'FilterFile'      , ValueFromPipelineByPropertyName = $true, Position = 2)]
         [Switch]
         $File
     )
@@ -23,12 +29,9 @@ function Get-nxChildItem
     {
         $lsParams  = @('-Al','--full-time','--group-directories-first')
 
-        switch ($PSboundParameters.keys)
+        if ($PSBoundParameters.ContainsKey('Recurse') -and $PSboundParameters['Recurse'])
         {
-            'Recurse'   { $lsParams += '--recursive' }
-            'Directory' { }
-            'File'      { }
-            default     { Write-Debug -Message "Parameter '$_' not added automatically." }
+            $lsParams += '-R' # Alpine linux does not support --recursive
         }
     }
 
@@ -37,7 +40,22 @@ function Get-nxChildItem
         foreach ($pathItem in $Path.Where{$_})
         {
             $pathItem = [System.IO.Path]::GetFullPath($pathItem, $PWD.Path)
-            Invoke-NativeCommand -Executable 'ls' -Parameters ($lsParams + @($pathItem)) | Convert-nxLsEntryToFileSystemInfo -InitialPath $pathItem
+            $unfilteredListCommand = {
+                Invoke-NativeCommand -Executable 'ls' -Parameters ($lsParams + @($pathItem)) | Convert-nxLsEntryToFileSystemInfo -InitialPath $pathItem
+            }
+
+            if ($PSCmdlet.ParameterSetName -eq 'FilterFile' -and $PSBoundParameters['File'])
+            {
+                &$unfilteredListCommand | Where-Object -FilterScript { $_.nxFileSystemItemType -eq [nxFileSystemItemType]::File }
+            }
+            elseif ($PSCmdlet.ParameterSetName -eq 'FilterDirectory' -and $PSBoundParameters['Directory'])
+            {
+                &$unfilteredListCommand | Where-Object -FilterScript { $_.nxFileSystemItemType -eq [nxFileSystemItemType]::Directory }
+            }
+            else
+            {
+                &$unfilteredListCommand
+            }
         }
     }
 }